Struct google_chat1::SpaceMethods

pub struct SpaceMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on space resources. It is not used directly, but through the HangoutsChat hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_chat1 as chat1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use chat1::HangoutsChat;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = HangoutsChat::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `list(...)`, `members_get(...)`, `members_list(...)`, `messages_create(...)`, `messages_delete(...)`, `messages_get(...)` and `messages_update(...)`
// to build up your call.
let rb = hub.spaces();

Methods

impl<'a, C, A> SpaceMethods<'a, C, A>

pub fn list(&self) -> SpaceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists spaces the caller is a member of.

pub fn messages_get(&self, name: &str) -> SpaceMessageGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a message.

Arguments

name - Required. Resource name of the message to be retrieved, in the form "spaces//messages/". Example: spaces/AAAAMpdlehY/messages/UMxbHmzDlr4.UMxbHmzDlr4

pub fn get(&self, name: &str) -> SpaceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a space.

Arguments

name - Required. Resource name of the space, in the form "spaces/*". Example: spaces/AAAAMpdlehY

pub fn messages_update(
&self,
request: Message,
name: &str
) -> SpaceMessageUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a message.

Arguments

request - No description provided.

name - Resource name, in the form "spaces//messages/". Example: spaces/AAAAMpdlehY/messages/UMxbHmzDlr4.UMxbHmzDlr4

pub fn members_list(&self, parent: &str) -> SpaceMemberListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists human memberships in a space.

Arguments

parent - Required. The resource name of the space for which membership list is to be fetched, in the form "spaces/*". Example: spaces/AAAAMpdlehY

pub fn members_get(&self, name: &str) -> SpaceMemberGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a membership.

Arguments

name - Required. Resource name of the membership to be retrieved, in the form "spaces//members/". Example: spaces/AAAAMpdlehY/members/105115627578887013105

pub fn messages_delete(&self, name: &str) -> SpaceMessageDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a message.

Arguments

name - Required. Resource name of the message to be deleted, in the form "spaces//messages/" Example: spaces/AAAAMpdlehY/messages/UMxbHmzDlr4.UMxbHmzDlr4

pub fn messages_create(
&self,
request: Message,
parent: &str
) -> SpaceMessageCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a message.

Arguments

request - No description provided.

parent - Required. Space resource name, in the form "spaces/*". Example: spaces/AAAAMpdlehY